# Make partitions type-stable