@mschauer 2018-10-09
Pretty straight forward to resolve the conflicts. I am unsure about
https://github.com/JuliaMath/Combinatorics.jl/pull/35/files#diff-d79e2641c7fdd050527c618f025993f3R249