@mschauer 2018-10-15
This is fine with https://github.com/ti-s/Combinatorics.jl/pull/1